:art: seperated POST endpoint from basic route
